# docs/swagger.yaml
basePath: /api/
definitions:
  http.AddTaskMemberRequest:
    properties:
      user_id:
        type: integer
    type: object
...
  httpErrors.RestError:
    properties:
      error:
        type: string
      status:
        type: integer
    type: object
...
  models.Task:
    properties:
      description:
        maxLength: 256
        type: string
      finished:
        type: boolean
      id:
        type: integer
      name:
        maxLength: 64
        type: string
      project_id:
        type: integer
    type: object
  models.User:
    properties:
      address:
        maxLength: 100
        type: string
      admin:
        type: boolean
      email:
        maxLength: 60
        type: string
      id:
        type: integer
      name:
        minLength: 2
        type: string
      password:
        minLength: 6
        type: string
      patronymic:
        type: string
      surname:
        maxLength: 60
        minLength: 2
        type: string
    required:
    - address
    - name
    - password
    - surname
    type: object
...
  utils.Response:
    properties:
      ok:
        type: boolean
    type: object
...
host: localhost:80
info:
  contact:
    email: armangokka@gmail.com
    name: API Support
    url: https://t.me/armanokka
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Time tracker REST API
  version: "1.0"
paths:
...
  /projects/{project_id}/tasks:
    get:
      description: Get project tasks
      parameters:
      - description: project id
        in: path
        name: project_id
        required: true
        type: string
      - description: Token that you get after authorization/registration
        in: header
        name: X-Access-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpErrors.RestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Get project tasks
      tags:
      - tasks
    post:
      description: Create project task
      parameters:
      - description: project id
        in: path
        name: project_id
        required: true
        type: string
      - description: task to be created
        in: body
        name: taskBody
        required: true
        schema:
          $ref: '#/definitions/models.Task'
      - description: Token that you get after authorization/registration
        in: header
        name: X-Access-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpErrors.RestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Create project task
      tags:
      - tasks
  /projects/{project_id}/tasks/{task_id}:
    delete:
      description: Delete project task
      parameters:
      - description: project id
        in: path
        name: project_id
        required: true
        type: string
      - description: task id
        in: path
        name: task_id
        required: true
        type: string
      - description: Token that you get after authorization/registration
        in: header
        name: X-Access-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpErrors.RestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Delete project task
      tags:
      - tasks
    patch:
      description: Update project task
      parameters:
      - description: project id
        in: path
        name: project_id
        required: true
        type: string
      - description: task id
        in: path
        name: task_id
        required: true
        type: string
      - description: updates to the task
        in: body
        name: updatedTaskBody
        required: true
        schema:
          $ref: '#/definitions/models.Task'
      - description: Token that you get after authorization/registration
        in: header
        name: X-Access-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpErrors.RestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Update project task
      tags:
      - tasks
  /projects/{project_id}/tasks/{task_id}/start:
    post:
      description: Start doing project task
      parameters:
      - description: project id
        in: path
        name: project_id
        required: true
        type: string
      - description: task id
        in: path
        name: task_id
        required: true
        type: string
      - description: Token that you get after authorization/registration
        in: header
        name: X-Access-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpErrors.RestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Start doing project task
      tags:
      - tasks
  /projects/{project_id}/tasks/{task_id}/stop:
    post:
      description: Stop doing project task
      parameters:
      - description: project id
        in: path
        name: project_id
        required: true
        type: string
      - description: task id
        in: path
        name: task_id
        required: true
        type: string
      - description: Token that you get after authorization/registration
        in: header
        name: X-Access-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpErrors.RestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Stop doing project task
      tags:
      - tasks
  /projects/{project_id}/tasks/{task_id}/users:
    get:
      description: Get task executors
      parameters:
      - description: project id
        in: path
        name: project_id
        required: true
        type: string
      - description: task id
        in: path
        name: task_id
        required: true
        type: string
      - description: Token that you get after authorization/registration
        in: header
        name: X-Access-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpErrors.RestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Get task executors
      tags:
      - tasks
    post:
      description: Add task executor
      parameters:
      - description: project id
        in: path
        name: project_id
        required: true
        type: string
      - description: task id
        in: path
        name: task_id
        required: true
        type: string
      - description: Token that you get after authorization/registration
        in: header
        name: X-Access-Token
        required: true
        type: string
      - description: add executor request body
        in: body
        name: addExecutorRequestBody
        required: true
        schema:
          $ref: '#/definitions/http.AddTaskMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpErrors.RestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Add task executor
      tags:
      - tasks
  /projects/{project_id}/tasks/{task_id}/users/{user_id}:
    delete:
      description: Remove task executor
      parameters:
      - description: project id
        in: path
        name: project_id
        required: true
        type: string
      - description: task id
        in: path
        name: task_id
        required: true
        type: string
      - description: executor id to be removed
        in: path
        name: user_id
        required: true
        type: string
      - description: Token that you get after authorization/registration
        in: header
        name: X-Access-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpErrors.RestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Remove task executor
      tags:
      - tasks
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"
...